requirements.txt
pwntools
shutil
selenium